commit, se agrego validacion de rang de fechas de evaluacion existente dentro del periodo

+++ Application.Main/Services/Employee/CollaboratorService.cs
@@ -110,12 +110,13 @@
             var evaluationCollaboratorId = await _unitOfWorkApp.Repository.EvaluationCollaboratorRepository
                     .Find(f => f.CollaboratorId.Equals(id) && (currentDate >= f.Evaluation.StartDate && currentDate <= f.Evaluation.EndDate))
                     .Select(s => s.Id)
-                    .FirstAsync();
+                    .FirstOrDefaultAsync();
 
 
             await _unitOfWorkApp.SaveChangesAsync();
 
-            await _service.DeleteAsync(evaluationCollaboratorId);
+            if(!evaluationCollaboratorId.Equals(Guid.Empty))
+                await _service.DeleteAsync(evaluationCollaboratorId);
 
             return true;
         }
